DenseNet/requirements.txt
torch~=1.11.0
pandas~=1.3.5
numpy~=1.21.5
matplotlib~=3.5.2
torchvision~=0.12.0
Pillow~=9.0.1
tqdm~=4.64.0
sklearn~=0.0
scikit-learn~=1.0.2
opencv-python~=4.5.5.64